lib: fix again memcmp export issues in the arch ksyms only
The last memcmp fix was not enough as Ralf Rösch just reported that it
did not fix build issues caused by the first one on mips (redefined
symbol). The proper fix finally consists in only exporting the symbol
in the affected architecture, i386. The final code is closer to what
it was prior to the first change.

If other architectures experience similar issues as i386 did, we'll add
the faulty symbols to their own export list. A cleaner solution would
have consisted in always exporting the symbols in string.c regardless
of the architecture as is done in 2.6, but this could have broken some
existing external patches, which is not desirable at this stage.